Split out ChunkedArray code into separate header and compilation unit
Split apart table.h / table_test.cc
