+++ cpp/src/arrow/compute/kernels/aggregate_test.cc
@@ -23,10 +23,10 @@
 #include <gtest/gtest.h>
 
 #include "arrow/array.h"
+#include "arrow/chunked_array.h"
 #include "arrow/compute/api_aggregate.h"
 #include "arrow/compute/kernels/aggregate_internal.h"
 #include "arrow/compute/kernels/test_util.h"
-#include "arrow/table.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
 #include "arrow/util/checked_cast.h"
