+++ cpp/src/arrow/compute/kernels/scalar_boolean_test.cc
@@ -22,9 +22,9 @@
 
 #include <gtest/gtest.h>
 
+#include "arrow/chunked_array.h"
 #include "arrow/compute/api_scalar.h"
 #include "arrow/compute/kernels/test_util.h"
-#include "arrow/table.h"
 #include "arrow/testing/gtest_common.h"
 #include "arrow/testing/gtest_util.h"
