+++ cpp/src/arrow/csv/column_builder.cc
@@ -26,6 +26,7 @@
 
 #include "arrow/array.h"
 #include "arrow/builder.h"
+#include "arrow/chunked_array.h"
 #include "arrow/csv/column_builder.h"
 #include "arrow/csv/converter.h"
 #include "arrow/csv/inference_internal.h"
@@ -33,7 +34,6 @@
 #include "arrow/csv/parser.h"
 #include "arrow/memory_pool.h"
 #include "arrow/status.h"
-#include "arrow/table.h"
 #include "arrow/type.h"
 #include "arrow/util/logging.h"
 #include "arrow/util/task_group.h"
