+++ cpp/src/arrow/stl.h
@@ -29,6 +29,7 @@
 
 #include "arrow/array.h"
 #include "arrow/builder.h"
+#include "arrow/chunked_array.h"
 #include "arrow/compute/api.h"
 #include "arrow/memory_pool.h"
 #include "arrow/status.h"
