+++ cpp/src/arrow/table_builder.h
@@ -22,8 +22,6 @@
 #include <vector>
 
 #include "arrow/array/builder_base.h"
-#include "arrow/memory_pool.h"
-#include "arrow/record_batch.h"
 #include "arrow/status.h"
 #include "arrow/type.h"
 #include "arrow/util/checked_cast.h"
@@ -31,6 +29,9 @@
 #include "arrow/util/visibility.h"
 
 namespace arrow {
+
+class MemoryPool;
+class RecordBatch;
 
 /// \class RecordBatchBuilder
 /// \brief Helper class for creating record batches iteratively given a known
