+++ cpp/src/arrow/table_test.cc
@@ -22,195 +22,19 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
+#include "arrow/array/array_base.h"
+#include "arrow/array/data.h"
 #include "arrow/array/util.h"
+#include "arrow/chunked_array.h"
 #include "arrow/record_batch.h"
 #include "arrow/status.h"
 #include "arrow/table.h"
 #include "arrow/testing/gtest_common.h"
 #include "arrow/testing/gtest_util.h"
-#include "arrow/testing/random.h"
 #include "arrow/type.h"
-#include "arrow/util/bit_util.h"
 #include "arrow/util/key_value_metadata.h"
 
 namespace arrow {
-
-class TestChunkedArray : public TestBase {
- protected:
-  virtual void Construct() {
-    one_ = std::make_shared<ChunkedArray>(arrays_one_);
-    if (!arrays_another_.empty()) {
-      another_ = std::make_shared<ChunkedArray>(arrays_another_);
-    }
-  }
-
-  ArrayVector arrays_one_;
-  ArrayVector arrays_another_;
-
-  std::shared_ptr<ChunkedArray> one_;
-  std::shared_ptr<ChunkedArray> another_;
-};
-
-TEST_F(TestChunkedArray, BasicEquals) {
-  std::vector<bool> null_bitmap(100, true);
-  std::vector<int32_t> data(100, 1);
-  std::shared_ptr<Array> array;
-  ArrayFromVector<Int32Type, int32_t>(null_bitmap, data, &array);
-  arrays_one_.push_back(array);
-  arrays_another_.push_back(array);
-
-  Construct();
-  ASSERT_TRUE(one_->Equals(one_));
-  ASSERT_FALSE(one_->Equals(nullptr));
-  ASSERT_TRUE(one_->Equals(another_));
-  ASSERT_TRUE(one_->Equals(*another_.get()));
-}
-
-TEST_F(TestChunkedArray, EqualsDifferingTypes) {
-  std::vector<bool> null_bitmap(100, true);
-  std::vector<int32_t> data32(100, 1);
-  std::vector<int64_t> data64(100, 1);
-  std::shared_ptr<Array> array;
-  ArrayFromVector<Int32Type, int32_t>(null_bitmap, data32, &array);
-  arrays_one_.push_back(array);
-  ArrayFromVector<Int64Type, int64_t>(null_bitmap, data64, &array);
-  arrays_another_.push_back(array);
-
-  Construct();
-  ASSERT_FALSE(one_->Equals(another_));
-  ASSERT_FALSE(one_->Equals(*another_.get()));
-}
-
-TEST_F(TestChunkedArray, EqualsDifferingLengths) {
-  std::vector<bool> null_bitmap100(100, true);
-  std::vector<bool> null_bitmap101(101, true);
-  std::vector<int32_t> data100(100, 1);
-  std::vector<int32_t> data101(101, 1);
-  std::shared_ptr<Array> array;
-  ArrayFromVector<Int32Type, int32_t>(null_bitmap100, data100, &array);
-  arrays_one_.push_back(array);
-  ArrayFromVector<Int32Type, int32_t>(null_bitmap101, data101, &array);
-  arrays_another_.push_back(array);
-
-  Construct();
-  ASSERT_FALSE(one_->Equals(another_));
-  ASSERT_FALSE(one_->Equals(*another_.get()));
-
-  std::vector<bool> null_bitmap1(1, true);
-  std::vector<int32_t> data1(1, 1);
-  ArrayFromVector<Int32Type, int32_t>(null_bitmap1, data1, &array);
-  arrays_one_.push_back(array);
-
-  Construct();
-  ASSERT_TRUE(one_->Equals(another_));
-  ASSERT_TRUE(one_->Equals(*another_.get()));
-}
-
-TEST_F(TestChunkedArray, EqualsDifferingMetadata) {
-  auto left_ty = list(field("item", int32()));
-
-  auto metadata = key_value_metadata({"foo"}, {"bar"});
-  auto right_ty = list(field("item", int32(), true, metadata));
-
-  std::vector<std::shared_ptr<Array>> left_chunks = {ArrayFromJSON(left_ty, "[[]]")};
-  std::vector<std::shared_ptr<Array>> right_chunks = {ArrayFromJSON(right_ty, "[[]]")};
-
-  ChunkedArray left(left_chunks);
-  ChunkedArray right(right_chunks);
-  ASSERT_TRUE(left.Equals(right));
-}
-
-TEST_F(TestChunkedArray, SliceEquals) {
-  arrays_one_.push_back(MakeRandomArray<Int32Array>(100));
-  arrays_one_.push_back(MakeRandomArray<Int32Array>(50));
-  arrays_one_.push_back(MakeRandomArray<Int32Array>(50));
-  Construct();
-
-  std::shared_ptr<ChunkedArray> slice = one_->Slice(125, 50);
-  ASSERT_EQ(slice->length(), 50);
-  AssertChunkedEqual(*one_->Slice(125, 50), *slice);
-
-  std::shared_ptr<ChunkedArray> slice2 = one_->Slice(75)->Slice(25)->Slice(25, 50);
-  ASSERT_EQ(slice2->length(), 50);
-  AssertChunkedEqual(*slice, *slice2);
-
-  // Making empty slices of a ChunkedArray
-  std::shared_ptr<ChunkedArray> slice3 = one_->Slice(one_->length(), 99);
-  ASSERT_EQ(slice3->length(), 0);
-  ASSERT_EQ(slice3->num_chunks(), 1);
-  ASSERT_TRUE(slice3->type()->Equals(one_->type()));
-
-  std::shared_ptr<ChunkedArray> slice4 = one_->Slice(10, 0);
-  ASSERT_EQ(slice4->length(), 0);
-  ASSERT_EQ(slice4->num_chunks(), 1);
-  ASSERT_TRUE(slice4->type()->Equals(one_->type()));
-
-  // Slicing an empty ChunkedArray
-  std::shared_ptr<ChunkedArray> slice5 = slice4->Slice(0, 10);
-  ASSERT_EQ(slice5->length(), 0);
-  ASSERT_EQ(slice5->num_chunks(), 1);
-  ASSERT_TRUE(slice5->type()->Equals(one_->type()));
-}
-
-TEST_F(TestChunkedArray, ZeroChunksIssues) {
-  ArrayVector empty = {};
-  auto no_chunks = std::make_shared<ChunkedArray>(empty, int8());
-
-  // ARROW-8911, assert that slicing is a no-op when there are zero-chunks
-  auto sliced = no_chunks->Slice(0, 0);
-  auto sliced2 = no_chunks->Slice(0, 5);
-  AssertChunkedEqual(*no_chunks, *sliced);
-  AssertChunkedEqual(*no_chunks, *sliced2);
-}
-
-TEST_F(TestChunkedArray, Validate) {
-  // Valid if empty
-  ArrayVector empty = {};
-  auto no_chunks = std::make_shared<ChunkedArray>(empty, utf8());
-  ASSERT_OK(no_chunks->ValidateFull());
-
-  random::RandomArrayGenerator gen(0);
-  arrays_one_.push_back(gen.Int32(50, 0, 100, 0.1));
-  Construct();
-  ASSERT_OK(one_->ValidateFull());
-
-  arrays_one_.push_back(gen.Int32(50, 0, 100, 0.1));
-  Construct();
-  ASSERT_OK(one_->ValidateFull());
-
-  arrays_one_.push_back(gen.String(50, 0, 10, 0.1));
-  Construct();
-  ASSERT_RAISES(Invalid, one_->ValidateFull());
-}
-
-TEST_F(TestChunkedArray, View) {
-  auto in_ty = int32();
-  auto out_ty = fixed_size_binary(4);
-#if ARROW_LITTLE_ENDIAN
-  auto arr = ArrayFromJSON(in_ty, "[2020568934, 2054316386, null]");
-  auto arr2 = ArrayFromJSON(in_ty, "[2020568934, 2054316386]");
-#else
-  auto arr = ArrayFromJSON(in_ty, "[1718579064, 1650553466, null]");
-  auto arr2 = ArrayFromJSON(in_ty, "[1718579064, 1650553466]");
-#endif
-  auto ex = ArrayFromJSON(out_ty, R"(["foox", "barz", null])");
-  auto ex2 = ArrayFromJSON(out_ty, R"(["foox", "barz"])");
-
-  ArrayVector chunks = {arr, arr2};
-  ArrayVector ex_chunks = {ex, ex2};
-  auto carr = std::make_shared<ChunkedArray>(chunks);
-  auto expected = std::make_shared<ChunkedArray>(ex_chunks);
-
-  ASSERT_OK_AND_ASSIGN(auto result, carr->View(out_ty));
-  AssertChunkedEqual(*expected, *result);
-
-  // Zero length
-  ArrayVector empty = {};
-  carr = std::make_shared<ChunkedArray>(empty, in_ty);
-  expected = std::make_shared<ChunkedArray>(empty, out_ty);
-  ASSERT_OK_AND_ASSIGN(result, carr->View(out_ty));
-  AssertChunkedEqual(*expected, *result);
-}
 
 class TestTable : public TestBase {
  public:
@@ -802,219 +626,6 @@ TEST_F(TestTable, AddColumn) {
   expected = Table::Make(
       ex_schema, {table.column(0), table.column(1), table.column(2), table.column(0)});
   ASSERT_TRUE(result->Equals(*expected));
-}
-
-class TestRecordBatch : public TestBase {};
-
-TEST_F(TestRecordBatch, Equals) {
-  const int length = 10;
-
-  auto f0 = field("f0", int32());
-  auto f1 = field("f1", uint8());
-  auto f2 = field("f2", int16());
-
-  auto metadata = key_value_metadata({"foo"}, {"bar"});
-
-  std::vector<std::shared_ptr<Field>> fields = {f0, f1, f2};
-  auto schema = ::arrow::schema({f0, f1, f2});
-  auto schema2 = ::arrow::schema({f0, f1});
-  auto schema3 = ::arrow::schema({f0, f1, f2}, metadata);
-
-  auto a0 = MakeRandomArray<Int32Array>(length);
-  auto a1 = MakeRandomArray<UInt8Array>(length);
-  auto a2 = MakeRandomArray<Int16Array>(length);
-
-  auto b1 = RecordBatch::Make(schema, length, {a0, a1, a2});
-  auto b2 = RecordBatch::Make(schema3, length, {a0, a1, a2});
-  auto b3 = RecordBatch::Make(schema2, length, {a0, a1});
-  auto b4 = RecordBatch::Make(schema, length, {a0, a1, a1});
-
-  ASSERT_TRUE(b1->Equals(*b1));
-  ASSERT_FALSE(b1->Equals(*b3));
-  ASSERT_FALSE(b1->Equals(*b4));
-
-  // Different metadata
-  ASSERT_TRUE(b1->Equals(*b2));
-  ASSERT_FALSE(b1->Equals(*b2, /*check_metadata=*/true));
-}
-
-TEST_F(TestRecordBatch, Validate) {
-  const int length = 10;
-
-  auto f0 = field("f0", int32());
-  auto f1 = field("f1", uint8());
-  auto f2 = field("f2", int16());
-
-  auto schema = ::arrow::schema({f0, f1, f2});
-
-  auto a0 = MakeRandomArray<Int32Array>(length);
-  auto a1 = MakeRandomArray<UInt8Array>(length);
-  auto a2 = MakeRandomArray<Int16Array>(length);
-  auto a3 = MakeRandomArray<Int16Array>(5);
-
-  auto b1 = RecordBatch::Make(schema, length, {a0, a1, a2});
-
-  ASSERT_OK(b1->ValidateFull());
-
-  // Length mismatch
-  auto b2 = RecordBatch::Make(schema, length, {a0, a1, a3});
-  ASSERT_RAISES(Invalid, b2->ValidateFull());
-
-  // Type mismatch
-  auto b3 = RecordBatch::Make(schema, length, {a0, a1, a0});
-  ASSERT_RAISES(Invalid, b3->ValidateFull());
-}
-
-TEST_F(TestRecordBatch, Slice) {
-  const int length = 7;
-
-  auto f0 = field("f0", int32());
-  auto f1 = field("f1", uint8());
-  auto f2 = field("f2", int8());
-
-  std::vector<std::shared_ptr<Field>> fields = {f0, f1, f2};
-  auto schema = ::arrow::schema(fields);
-
-  auto a0 = MakeRandomArray<Int32Array>(length);
-  auto a1 = MakeRandomArray<UInt8Array>(length);
-  auto a2 = ArrayFromJSON(int8(), "[0, 1, 2, 3, 4, 5, 6]");
-
-  auto batch = RecordBatch::Make(schema, length, {a0, a1, a2});
-
-  auto batch_slice = batch->Slice(2);
-  auto batch_slice2 = batch->Slice(1, 5);
-
-  ASSERT_EQ(batch_slice->num_rows(), batch->num_rows() - 2);
-
-  for (int i = 0; i < batch->num_columns(); ++i) {
-    ASSERT_EQ(2, batch_slice->column(i)->offset());
-    ASSERT_EQ(length - 2, batch_slice->column(i)->length());
-
-    ASSERT_EQ(1, batch_slice2->column(i)->offset());
-    ASSERT_EQ(5, batch_slice2->column(i)->length());
-  }
-
-  // ARROW-9143: RecordBatch::Slice was incorrectly setting a2's
-  // ArrayData::null_count to kUnknownNullCount
-  ASSERT_EQ(batch_slice->column(2)->data()->null_count, 0);
-  ASSERT_EQ(batch_slice2->column(2)->data()->null_count, 0);
-}
-
-TEST_F(TestRecordBatch, AddColumn) {
-  const int length = 10;
-
-  auto field1 = field("f1", int32());
-  auto field2 = field("f2", uint8());
-  auto field3 = field("f3", int16());
-
-  auto schema1 = ::arrow::schema({field1, field2});
-  auto schema2 = ::arrow::schema({field2, field3});
-  auto schema3 = ::arrow::schema({field2});
-
-  auto array1 = MakeRandomArray<Int32Array>(length);
-  auto array2 = MakeRandomArray<UInt8Array>(length);
-  auto array3 = MakeRandomArray<Int16Array>(length);
-
-  auto batch1 = RecordBatch::Make(schema1, length, {array1, array2});
-  auto batch2 = RecordBatch::Make(schema2, length, {array2, array3});
-  auto batch3 = RecordBatch::Make(schema3, length, {array2});
-
-  const RecordBatch& batch = *batch3;
-
-  // Negative tests with invalid index
-  ASSERT_RAISES(Invalid, batch.AddColumn(5, field1, array1));
-  ASSERT_RAISES(Invalid, batch.AddColumn(2, field1, array1));
-  ASSERT_RAISES(Invalid, batch.AddColumn(-1, field1, array1));
-
-  // Negative test with wrong length
-  auto longer_col = MakeRandomArray<Int32Array>(length + 1);
-  ASSERT_RAISES(Invalid, batch.AddColumn(0, field1, longer_col));
-
-  // Negative test with mismatch type
-  ASSERT_RAISES(Invalid, batch.AddColumn(0, field1, array2));
-
-  ASSERT_OK_AND_ASSIGN(auto new_batch, batch.AddColumn(0, field1, array1));
-  AssertBatchesEqual(*new_batch, *batch1);
-
-  ASSERT_OK_AND_ASSIGN(new_batch, batch.AddColumn(1, field3, array3));
-  AssertBatchesEqual(*new_batch, *batch2);
-
-  ASSERT_OK_AND_ASSIGN(auto new_batch2, batch.AddColumn(1, "f3", array3));
-  AssertBatchesEqual(*new_batch2, *new_batch);
-
-  ASSERT_TRUE(new_batch2->schema()->field(1)->nullable());
-}
-
-TEST_F(TestRecordBatch, RemoveColumn) {
-  const int length = 10;
-
-  auto field1 = field("f1", int32());
-  auto field2 = field("f2", uint8());
-  auto field3 = field("f3", int16());
-
-  auto schema1 = ::arrow::schema({field1, field2, field3});
-  auto schema2 = ::arrow::schema({field2, field3});
-  auto schema3 = ::arrow::schema({field1, field3});
-  auto schema4 = ::arrow::schema({field1, field2});
-
-  auto array1 = MakeRandomArray<Int32Array>(length);
-  auto array2 = MakeRandomArray<UInt8Array>(length);
-  auto array3 = MakeRandomArray<Int16Array>(length);
-
-  auto batch1 = RecordBatch::Make(schema1, length, {array1, array2, array3});
-  auto batch2 = RecordBatch::Make(schema2, length, {array2, array3});
-  auto batch3 = RecordBatch::Make(schema3, length, {array1, array3});
-  auto batch4 = RecordBatch::Make(schema4, length, {array1, array2});
-
-  const RecordBatch& batch = *batch1;
-  std::shared_ptr<RecordBatch> result;
-
-  // Negative tests with invalid index
-  ASSERT_RAISES(Invalid, batch.RemoveColumn(3));
-  ASSERT_RAISES(Invalid, batch.RemoveColumn(-1));
-
-  ASSERT_OK_AND_ASSIGN(auto new_batch, batch.RemoveColumn(0));
-  AssertBatchesEqual(*new_batch, *batch2);
-
-  ASSERT_OK_AND_ASSIGN(new_batch, batch.RemoveColumn(1));
-  AssertBatchesEqual(*new_batch, *batch3);
-
-  ASSERT_OK_AND_ASSIGN(new_batch, batch.RemoveColumn(2));
-  AssertBatchesEqual(*new_batch, *batch4);
-}
-
-TEST_F(TestRecordBatch, RemoveColumnEmpty) {
-  const int length = 10;
-
-  auto field1 = field("f1", int32());
-  auto schema1 = ::arrow::schema({field1});
-  auto array1 = MakeRandomArray<Int32Array>(length);
-  auto batch1 = RecordBatch::Make(schema1, length, {array1});
-
-  ASSERT_OK_AND_ASSIGN(auto empty, batch1->RemoveColumn(0));
-  ASSERT_EQ(batch1->num_rows(), empty->num_rows());
-
-  ASSERT_OK_AND_ASSIGN(auto added, empty->AddColumn(0, field1, array1));
-  AssertBatchesEqual(*added, *batch1);
-}
-
-TEST_F(TestRecordBatch, ToFromEmptyStructArray) {
-  auto batch1 =
-      RecordBatch::Make(::arrow::schema({}), 10, std::vector<std::shared_ptr<Array>>{});
-  ASSERT_OK_AND_ASSIGN(auto struct_array, batch1->ToStructArray());
-  ASSERT_EQ(10, struct_array->length());
-  ASSERT_OK_AND_ASSIGN(auto batch2, RecordBatch::FromStructArray(struct_array));
-  ASSERT_TRUE(batch1->Equals(*batch2));
-}
-
-TEST_F(TestRecordBatch, FromStructArrayInvalidType) {
-  ASSERT_RAISES(Invalid, RecordBatch::FromStructArray(MakeRandomArray<Int32Array>(10)));
-}
-
-TEST_F(TestRecordBatch, FromStructArrayInvalidNullCount) {
-  auto struct_array =
-      ArrayFromJSON(struct_({field("f1", int32())}), R"([{"f1": 1}, null])");
-  ASSERT_RAISES(Invalid, RecordBatch::FromStructArray(struct_array));
 }
 
 class TestTableBatchReader : public TestBase {};
