+++ cpp/src/parquet/column_reader.cc
@@ -30,7 +30,7 @@
 
 #include "arrow/array.h"
 #include "arrow/builder.h"
-#include "arrow/table.h"
+#include "arrow/chunked_array.h"
 #include "arrow/type.h"
 #include "arrow/util/bit_stream_utils.h"
 #include "arrow/util/checked_cast.h"
